restructured the functionality related to drone drag-and-drop and sadezone drawing into separate components, and included the capability to edit the safe zone from its bottom right corner.
