+++ frontend/src/components/cesium/DrawSadeZone.jsx
@@ -1,0 +1,84 @@
+import React, { useRef, useEffect, useState } from 'react';
+import { Entity } from 'resium';
+import { ScreenSpaceEventType, Cartographic, Color, Rectangle, KeyboardEventModifier, 
+    ScreenSpaceEventHandler, Ellipsoid, defined } from 'cesium';
+import PropTypes from 'prop-types';
+
+const DrawSadeZone = ({ viewerReady, viewerRef, setNewCameraPosition }) => {
+  const [mouseDown, setMouseDown] = useState(false);
+  const [firstPoint, setFirstPoint] = useState(null);
+  const [rectangleCoordinates, setRectangleCoordinates] = useState(null);
+  const [showSelector, setShowSelector] = useState(false);
+  const [isEditing, setIsEditing] = useState(false);
+  const [selectedPoint, setSelectedPoint] = useState(null);
+  const [corners, setCorners] = useState({});
+  const [selectedCorner, setSelectedCorner] = useState(null);
+
+  useEffect(() => {
+    if (!viewerReady) return;
+    const viewer = viewerRef.current.cesiumElement;
+    const handler = new ScreenSpaceEventHandler(viewer.scene.canvas);
+
+    // Dragging logic
+    handler.setInputAction((movement) => {
+      if (!mouseDown) return;
+      const cartesian = viewer.camera.pickEllipsoid(movement.endPosition, viewer.scene.globe.ellipsoid);
+      if (cartesian) {
+        const tempCartographic = Cartographic.fromCartesian(cartesian, Ellipsoid.WGS84);
+        if (!firstPoint) {
+          setFirstPoint(tempCartographic);
+        } else {
+          const newRectangle = new Rectangle(
+            Math.min(tempCartographic.longitude, firstPoint.longitude),
+            Math.min(tempCartographic.latitude, firstPoint.latitude),
+            Math.max(tempCartographic.longitude, firstPoint.longitude),
+            Math.max(tempCartographic.latitude, firstPoint.latitude)
+          );
+          setRectangleCoordinates(newRectangle);
+          setShowSelector(true);
+          setNewCameraPosition();
+        }
+      }
+    }, ScreenSpaceEventType.MOUSE_MOVE, KeyboardEventModifier.SHIFT);
+
+    handler.setInputAction(() => {
+      setMouseDown(true);
+    }, ScreenSpaceEventType.LEFT_DOWN, KeyboardEventModifier.SHIFT);
+
+    handler.setInputAction(() => {
+      setMouseDown(false);
+      setFirstPoint(null);
+      setShowSelector(false);
+      setNewCameraPosition();
+    }, ScreenSpaceEventType.LEFT_UP, KeyboardEventModifier.SHIFT);
+
+    return () => {
+      handler.destroy();
+    };
+  }, [viewerReady, mouseDown, firstPoint, rectangleCoordinates, showSelector]);
+
+  return (
+    <>
+    {showSelector &&
+    <Entity
+    rectangle={{
+    coordinates: rectangleCoordinates,
+    material: Color.GREEN.withAlpha(0.5),
+    outline: true,
+    outlineColor: Color.WHITE,
+    outlineWidth: 2,
+    extrudedHeight: 30.0,
+    }}
+    />
+    }
+    </>
+  );
+};
+
+DrawSadeZone.propTypes = {
+    viewerReady: PropTypes.bool.isRequired,
+    viewerRef: PropTypes.object.isRequired,
+    setNewCameraPosition: PropTypes.func.isRequired
+};
+
+export default DrawSadeZone;
